+++ route/src/main/java/sparta/rooibos/route/application/dto/request/route/GetOptimizedRouteRequest.java
@@ -1,4 +1,4 @@
-package sparta.rooibos.route.application.dto.request;
+package sparta.rooibos.route.application.dto.request.route;
 
 import sparta.rooibos.route.application.core.DijkstraAlgorithm;
